+++ types/src/state_store/errors.rs
@@ -5,7 +5,7 @@
 use thiserror::Error;
 
 #[derive(Debug, Error)]
-pub enum StateviewError {
+pub enum StateViewError {
     #[error("{0} not found.")]
     NotFound(String),
     /// Other non-classified error.
@@ -15,7 +15,7 @@ pub enum StateviewError {
     BcsError(#[from] bcs::Error),
 }
 
-impl From<anyhow::Error> for StateviewError {
+impl From<anyhow::Error> for StateViewError {
     fn from(error: anyhow::Error) -> Self {
         Self::Other(format!("{}", error))
     }